Specify Mac OS version of libfftw3 and update README
When installing the `fftw` using brew into `/usr/local/lib`, the
Makefile located in `src` must be updated to reference the `libfftw3`
library found in `/usr/local/lib` to be able to compile EnKF_seir.

In addition, the README has been updated to specify respective commands
for Linux and Mac OS, with a special note in the `Mac` section of `Run
the Project`, to add a symlink to enable users to run `seir` from
anywhere on their Mac file system.

Lastly, other syntactical updates have been made to the README for
consistency.